requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
ultralytics
streamlit-webrtc>=0.47.0
opencv-python-headless
av
numpy
